requirements.txt
asgiref==3.8.1
Django==5.0.7
django-allauth==64.0.0
pillow==10.4.0
sorl-thumbnail==12.10.0
sqlparse==0.5.1
tzdata==2024.1
